fix: 공연금액 타입 변경

+++ src/main/java/com/profect/tickle/domain/performance/entity/Performance.java
@@ -41,7 +41,7 @@ public class Performance {
     private String title;
 
     @Column(name = "performance_price", nullable = false)
-    private Short price;
+    private Integer price;
 
     @Column(name = "performance_date", nullable = false)
     private LocalDateTime date;
